pyproject.toml
[project]
name = "cognito-auth"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "django>=4.2.7",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt[crypto]>=5.3.0",
    "pycognito>=2023.5.0",
    "boto3>=1.29.1",
    "django-environ>=0.11.2",
    "setuptools>=68.2.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "ruff>=0.1.5",
    "isort>=5.12.0",
    "black>=23.11.0",
    "mypy>=1.7.0",
    "flake8>=5.0.4",
    "pylint>=3.0.2",
    "autopep8>=2.0.4",
]

[tool.hatch.metadata]
allow-direct-references = true
